+++ packages/core/types/index.d.ts
@@ -1,6 +1,6 @@
 export { AddonRegister, AddonGet, Addon, Subscriber, SideEffect, UnsubscribeFn, Cache, RawParams, Params } from './interface';
 export { Path } from './utils/createPath';
-export { Route, RouteDescriptor, Title, ParamParser, ParamParsers, LoadModifiers, LoadFn, PreloadFn } from './utils/createRoute';
+export { Route, RouteDescriptor, Title, ParamParser, ParamParsers, LoadRoute, LoadModifiers, LoadFn, PreloadFn } from './utils/createRoute';
 export { Response, RedirectResponse, AnyResponse } from './utils/createResponse';
 export { CuriConfig, ConfigOptions } from './createConfig';
 import createConfig from './createConfig';